Refactor project exceptions structure, add excludes for generated code coverage
