+++ bot/src/main/java/edu/java/bot/exception/ChatIdNotFoundException.java
@@ -1,4 +1,4 @@
-package edu.java.bot.controller.exception;
+package edu.java.bot.exception;
 
 public class ChatIdNotFoundException extends RuntimeException {
     public ChatIdNotFoundException(String message) {
